requirements.txt
ipykernel
pandas
numpy
matplotlib
plotly
seaborn
scikit-learn
pymongo
from_root
dill
certifi
PyYAML
boto3
mypy-boto3-s3
botocore
fastapi
python-multipart
uvicorn
setuptools
imblearn
nltk
mlflow
xgboost
pre-commit
pytest
httpx
-e .
